refactor: try to connect with /api/search
